Move workspaces list to own file
